// src/api-serverless/src/generated/models/CreateNewProfileProxy.ts
/**
 * Seize API
 * Following is the API interface description for most common operations in Seize API. Some modifying endpoints may need authentication token.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateNewProfileProxy {
  'target_id': string;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'target_id',
      baseName: 'target_id',
      type: 'string',
      format: ''
    }
  ];

  static getAttributeTypeMap() {
    return CreateNewProfileProxy.attributeTypeMap;
  }

  public constructor() {}
}