// src/api-serverless/src/generated/models/CreateNewProfileProxyActionType.ts
/**
 * Seize API
 * Following is the API interface description for most common operations in Seize API. Some modifying endpoints may need authentication token.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export enum CreateNewProfileProxyActionType {
  AllocateRep = 'ALLOCATE_REP',
  AllocateCic = 'ALLOCATE_CIC',
  CreateWave = 'CREATE_WAVE',
  ReadWave = 'READ_WAVE',
  CreateDropToWave = 'CREATE_DROP_TO_WAVE',
  RateWaveDrop = 'RATE_WAVE_DROP'
}